fix part2
