+++ lib/screens/verifikasi_input.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:finlog/styles/colors.dart'; // Assuming this file has the necessary colors
 import 'package:finlog/screens/bill_details_screen.dart';
-import 'package:finlog/screens/text_input_page/journal_chat_input_screen.dart';
+import 'package:finlog/screens/text_input_page/journal_entry_input_screen.dart';
 import 'package:finlog/screens/text_input_page/manual_input_screen.dart'; // Import ManualInputScreen
 import 'package:finlog/services/gemini_service.dart';
 import 'package:finlog/models/manual_input_data.dart'; // Import ManualInputData
@@ -116,7 +116,7 @@ class _VerifikasiInputScreenState extends State<VerifikasiInputScreen> {
       Navigator.pushReplacement(
         context,
         MaterialPageRoute(
-          builder: (context) => JournalChatInputScreen(
+          builder: (context) => JournalEntryInputScreen(
             selectedDate: widget.journalDate ?? DateTime.now(), // Pass the original journalDate back, or today's date as fallback
           ),
         ),
